switch to zero-value timesource
