+++ timeseries/blueflood/blueflood_single_test.go
@@ -29,10 +29,10 @@ import (
 func TestBluefloodHTTPQueriesSingle(t *testing.T) {
 	// from -2m to now
 	nowMillis := int64(739908000000)
-	nowFunc := func() time.Time {
+	nowFunc := TimeSource{GetTime: func() time.Time {
 		timeValue := time.Unix(nowMillis/1000, nowMillis%1000*1e6)
 		return timeValue
-	}
+	}}
 	makeRange := func(beforeStart time.Duration, beforeEnd time.Duration, resolution time.Duration) api.Timerange {
 		if beforeStart < beforeEnd {
 			t.Fatalf("Before start must be at least as large as before end.")
@@ -141,10 +141,10 @@ func TestBluefloodHTTPQueriesSingle(t *testing.T) {
 func TestBluefloodHTTPQueriesMultiResolutionSingle(t *testing.T) {
 	// from -30d5h to -14d17h
 	nowMillis := int64(739908000000)
-	nowFunc := func() time.Time {
+	nowFunc := TimeSource{GetTime: func() time.Time {
 		timeValue := time.Unix(nowMillis/1000, nowMillis%1000*1e6)
 		return timeValue
-	}
+	}}
 	offset30 := int64(30 * 24 * 60 * 60 * 1000) // 30 days, TTL on 5m resolution
 	offset15 := int64(15 * 24 * 60 * 60 * 1000) // 15 days, first available on 60
 	makeRange := func(beforeStart time.Duration, beforeEnd time.Duration, resolution time.Duration) api.Timerange {
